pyproject.toml
[project]
name = "lam4inv-openai"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "httpx[socks]>=0.28.1",
  "openai>=1.0.0",
  "python-dotenv>=1.0.1",
  "z3-solver>=4.13.0.0",
]

[tool.uv]
package = false
